refactor(statistics): create statistics module
