--- backend/src/modules/organization/dto/organization-request-filter.dto.ts
@@ -1,8 +1,0 @@
-import { IsEnum } from 'class-validator';
-import { OrganizationRequestFilter } from '../enums/organization-request-statistics.enum';
-
-export class OrganizationStatisticsFilterDto {
-  @IsEnum(OrganizationRequestFilter)
-  organizationRequestFilter: OrganizationRequestFilter =
-    OrganizationRequestFilter._30_DAYS;
-}
