+++ backend/src/modules/organization/repositories/index.ts
@@ -8,4 +8,3 @@ export * from './contact.repository';
 export * from './partner.repository';
 export * from './investor.repository';
 export * from './organization-view.repository';
-export * from './organization-status-statistics-view.repository';
