+++ backend/src/modules/statistics/enums/organization-statistics-type.enum.ts
