+++ backend/src/modules/statistics/repositories/organization-status-statistics-view.repository.ts
